Olenas branch5

+++ src/member/MemberService.java
@@ -2,6 +2,8 @@ package member;
 
 import base.BaseRepository;
 import base.BaseService;
+import loan.LoanService;
+import loan.LoanSummaryDto;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -11,6 +13,7 @@ import java.util.Optional;
 public class MemberService extends BaseService<Member, Integer> {
 
     private final MemberRepository memberRepository = new MemberRepository();
+    private final LoanService loanService = new LoanService();
     private final MemberMapper memberMapper = new MemberMapper();
 
     // =========================================================
@@ -28,41 +31,14 @@ public class MemberService extends BaseService<Member, Integer> {
         MemberValidator.validateId(id);
     }
 
-
-    public Optional<Member> getByEmail(String email) throws SQLException {
-        MemberValidator.validateEmail(email);
-        return memberRepository.getByEmail(email);
-    }
-
-
-
-    // =========================================================
-    // Reader actions - 3. View my profile
-    //                        --> 1. View my profile
-    // =========================================================
-    // 1 showCurrentMemberProfile (MemberController)
-    //    -> 2 getProfileById (MemberService) +
-    //       -> 3 validateId (MemberService) +
-    //          -> 4 validateId (BaseService)
-    //       -> 5 getById (MemberRepository)
-    //       -> 6 toProfileDto (MemberMapper)
-    //    -> 7 printProfileMember (MemberController)
-    //    -> 8 printError (ConsolePrinter)
-
     public Optional<MemberProfileDto> getProfileById(Integer id) throws SQLException {
         validateId(id);
         return memberRepository.getById(id).map(memberMapper::toProfileDto);
     }
 
-
-
     //get
-    public Optional<MemberAdminDto> getByEmailForViewForAdmin(String email) throws SQLException {
-        MemberValidator.validateEmail(email);
-        return memberRepository.getByEmail(email)
-                .map(memberMapper::toAdminDto);
-    }
-    public List<MemberAdminDto> getAllForAdminView() throws SQLException {
+    public List<MemberAdminDto> getAllForAdminView(Member currentMember) throws SQLException {
+        validateLibrarianAccess(currentMember);
         List<Member> members = memberRepository.getAll();
         List<MemberAdminDto> result = new ArrayList<>();
 
@@ -72,7 +48,8 @@ public class MemberService extends BaseService<Member, Integer> {
 
         return result;
     }
-    public List<MemberAdminDto> searchMembersForAdmin(String keyword) throws SQLException {
+    public List<MemberAdminDto> searchMembersForAdmin(Member currentMember, String keyword) throws SQLException {
+        validateLibrarianAccess(currentMember);
         if (keyword == null || keyword.trim().isEmpty()) {
             throw new IllegalArgumentException("Search keyword cannot be empty.");
         }
@@ -88,7 +65,8 @@ public class MemberService extends BaseService<Member, Integer> {
     }
 
     //update
-    public Optional<MemberAdminDto> updateMemberByAdmin(UpdateMemberDto dto) throws SQLException {
+    public Optional<MemberAdminDto> updateMemberByAdmin(Member currentMember,UpdateMemberDto dto) throws SQLException {
+        validateLibrarianAccess(currentMember);
         MemberValidator.validateUpdateMemberAdminDto(dto);
 
         Optional<Member> optionalMember = memberRepository.getById(dto.getId());
@@ -163,8 +141,10 @@ public class MemberService extends BaseService<Member, Integer> {
         memberRepository.updatePassword(dto.getMemberId(), dto.getNewPassword());
         return true;
     }
+
     //create
-    public Optional<MemberAdminDto> createMemberByAdmin(CreateMemberDto dto) throws SQLException{
+    public Optional<MemberAdminDto> createMemberByAdmin(Member currentMember, CreateMemberDto dto) throws SQLException{
+        validateLibrarianAccess(currentMember);
         MemberValidator.validateCreateMemberDto(dto);
         Optional<Member> optionalMember = memberRepository.getByEmail(dto.getEmail());
         if (optionalMember.isPresent()){
@@ -180,20 +160,26 @@ public class MemberService extends BaseService<Member, Integer> {
     }
 
     //delete
-
-    public boolean deleteMemberByAdmin(Integer memberId) throws SQLException {
+    public boolean deleteMemberByAdmin(Member currentMember, Integer memberId) throws SQLException {
+        validateLibrarianAccess(currentMember);
         MemberValidator.validateId(memberId);
+
 
         Optional<Member> optionalMember = memberRepository.getById(memberId);
         if (optionalMember.isEmpty()) {
             return false;
         }
 
+        List<LoanSummaryDto> activeLoans = loanService.getActiveLoansByMember(memberId);
+        if (!activeLoans.isEmpty()) {
+            throw new MemberException(
+                    "Member cannot be deleted because they have borrowed books that are not returned yet."
+            );
+        }
+
         memberRepository.deleteById(memberId);
         return true;
     }
-
-
 
 
 
@@ -222,15 +208,15 @@ public class MemberService extends BaseService<Member, Integer> {
 
     }
 
-    public void validateLibrarianAccess(Member currentMember) {
-        if (currentMember == null) {
-            throw new IllegalArgumentException("User is not authorized.");
-        }
-
-        if (!"LIBRARIAN".equalsIgnoreCase(currentMember.getRole())) {
-            throw new IllegalArgumentException("Access denied.");
-        }
-    }
+//    public void validateLibrarianAccess(Member currentMember) {
+//        if (currentMember == null) {
+//            throw new IllegalArgumentException("User is not authorized.");
+//        }
+//
+//        if (!"LIBRARIAN".equalsIgnoreCase(currentMember.getRole())) {
+//            throw new IllegalArgumentException("Access denied.");
+//        }
+//    }
 
     // =========================================================
     // Validation helpers
@@ -244,3 +230,17 @@ public class MemberService extends BaseService<Member, Integer> {
         }
     }
 }
+
+
+// =========================================================
+// Reader actions - 3. View my profile
+//                        --> 1. View my profile
+// =========================================================
+// 1 showCurrentMemberProfile (MemberController)
+//    -> 2 getProfileById (MemberService) +
+//       -> 3 validateId (MemberService) +
+//          -> 4 validateId (BaseService)
+//       -> 5 getById (MemberRepository)
+//       -> 6 toProfileDto (MemberMapper)
+//    -> 7 printProfileMember (MemberController)
+//    -> 8 printError (ConsolePrinter)
